Refactor launcher package to lifecycle
